Encapsulate access to the HTTP client to prepare for making requests replayable.
