+++ src/harvester/mod.rs
@@ -1,18 +1,16 @@
 pub mod ckan;
+pub mod client;
 pub mod csw;
 pub mod doris_bfs;
 pub mod geo_network_q;
 pub mod wasser_de;
 
 use std::fmt;
-use std::future::Future;
 use std::io::Read;
 
-use anyhow::{Error, Result};
-use bytes::Bytes;
+use anyhow::Result;
 use cap_std::fs::{Dir, OpenOptions as FsOpenOptions};
 use serde::Deserialize;
-use tokio::time::{sleep, Duration};
 use toml::from_str;
 use url::Url;
 
@@ -31,41 +29,6 @@ async fn write_dataset(dir: &Dir, id: String, dataset: Dataset) -> Result<()> {
     dataset.write(file).await?;
 
     Ok(())
-}
-
-trait Response {}
-
-impl Response for Bytes {}
-
-impl Response for String {}
-
-async fn with_retry<A, F, T, E>(mut action: A) -> Result<T>
-where
-    A: FnMut() -> F,
-    F: Future<Output = Result<T, E>>,
-    T: Response,
-    E: Into<Error> + fmt::Display,
-{
-    let mut attempts = 0;
-    let mut duration = Duration::from_secs(1);
-
-    loop {
-        match action().await {
-            Ok(val) => return Ok(val),
-            Err(err) => {
-                if attempts < 3 {
-                    tracing::warn!("Request failed but will be retried: {:#}", err);
-
-                    sleep(duration).await;
-
-                    attempts += 1;
-                    duration *= 10;
-                } else {
-                    return Err(err.into());
-                }
-            }
-        }
-    }
 }
 
 #[derive(Debug, Deserialize)]
@@ -147,60 +110,4 @@ pub enum Type {
     WasserDe,
     GeoNetworkQ,
     DorisBfs,
-}
-
-#[cfg(test)]
-mod tests {
-    use super::*;
-
-    use anyhow::anyhow;
-    use tokio::time::{pause, Instant};
-
-    #[tokio::test]
-    async fn with_retry_fowards_success() {
-        pause();
-        let start = Instant::now();
-
-        with_retry::<_, _, _, Error>(|| async { Ok(Bytes::new()) })
-            .await
-            .unwrap();
-
-        assert_eq!(start.elapsed().as_secs(), 0);
-    }
-
-    #[tokio::test]
-    async fn with_retry_fowards_failure() {
-        pause();
-        let start = Instant::now();
-
-        with_retry::<_, _, Bytes, _>(|| async { Err(anyhow!("failure")) })
-            .await
-            .unwrap_err();
-
-        assert_eq!(start.elapsed().as_secs(), 1 + 10 + 100);
-    }
-
-    #[tokio::test]
-    async fn with_retry_retries_three_times() {
-        pause();
-        let start = Instant::now();
-
-        let mut count = 0;
-
-        with_retry(|| {
-            count += 1;
-
-            async move {
-                if count > 3 {
-                    Ok(Bytes::new())
-                } else {
-                    Err(anyhow!("failure"))
-                }
-            }
-        })
-        .await
-        .unwrap();
-
-        assert_eq!(start.elapsed().as_secs(), 1 + 10 + 100);
-    }
 }
