+++ src/harvester/wasser_de.rs
@@ -1,12 +1,11 @@
 use anyhow::{anyhow, Result};
 use cap_std::fs::Dir;
-use reqwest::Client;
 use serde::{Deserialize, Serialize};
 use serde_json::from_slice;
 
 use crate::{
     dataset::{Dataset, License},
-    harvester::{with_retry, write_dataset, Source},
+    harvester::{client::Client, write_dataset, Source},
 };
 
 pub async fn harvest(dir: &Dir, client: &Client, source: &Source) -> Result<(usize, usize, usize)> {
@@ -14,17 +13,18 @@ pub async fn harvest(dir: &Dir, client: &Client, source: &Source) -> Result<(usi
         .url
         .join("rest/BaseController/FilterElements/V_REP_BASE_VALID")?;
 
-    let body = with_retry(|| async {
-        client
-            .post(url.clone())
-            .json(&Request { filter: Filter {} })
-            .send()
-            .await?
-            .error_for_status()?
-            .bytes()
-            .await
-    })
-    .await?;
+    let body = client
+        .make_request(|client| async {
+            client
+                .post(url.clone())
+                .json(&Request { filter: Filter {} })
+                .send()
+                .await?
+                .error_for_status()?
+                .bytes()
+                .await
+        })
+        .await?;
 
     let response = from_slice::<Response>(&body)?;
